# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk mengelola data bioskop",
        "title": "Bioskop API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/bioskop": {
            "get": {
                "description": "Mengambil semua data bioskop",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bioskop"
                ],
                "summary": "Daftar semua bioskop",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Bioskop"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Menambahkan bioskop ke database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bioskop"
                ],
                "summary": "Tambah bioskop baru",
                "parameters": [
                    {
                        "description": "Data bioskop",
                        "name": "bioskop",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Bioskop"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Bioskop"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/bioskop/{id}": {
            "get": {
                "description": "Mengambil detail bioskop berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bioskop"
                ],
                "summary": "Detail bioskop berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID bioskop",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bioskop"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbarui data bioskop berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bioskop"
                ],
                "summary": "Perbarui data bioskop",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID bioskop",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data bioskop baru",
                        "name": "bioskop",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Bioskop"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus bioskop berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bioskop"
                ],
                "summary": "Hapus bioskop",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID bioskop",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Bioskop": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lokasi": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        }
    }
}